requirements.txt
albumentations==0.4.3
colorama==0.4.5
cycler==0.11.0
decorator==4.4.2
imageio==2.15.0
imgaug==0.2.6
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
Pillow==8.4.0
pydicom==2.3.1
pyparsing==3.0.7
python-dateutil==2.9.0.post0
PyWavelets==1.1.1
PyYAML==6.0.1
scikit-image==0.17.2
scipy==1.5.4
six==1.17.0
tifffile==2020.9.3
tqdm==4.64.1
zipp==3.6.0
